Remove need for jupyterlab-myst for javelin throwing example
Now one can see the results of the calculations directly in the notebook
and one isn't reliant on mybinder.org in order to see these values
correctly.  If one wants to change the values, then mybinder.org is
necessary (or a local version of jupyterlab) but now we're able to
support clients that *can't* use mybinder.org (or not easily).