Fix numbers from automatic calculations in throwing a javelin exercise
I noticed that there were negative numbers coming out of some of the
calculations where they shouldn't be negative. This was because the
variables were being reused. The solution was to add subscripts to the
variables so that they uniquely specify the given situation and hence
will always perform the correct calculation now.